requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
pydantic[email]
sqladmin
jinja2
python-multipart
itsdangerous
bcrypt
slowapi
authlib
python-dotenv

# Testing
pytest
httpx
